+++ src/store/modules/lodging/state.js
@@ -1,0 +1,25 @@
+import { DataSet } from 'vue2vis';
+
+const state = {
+	message: '',
+	updatingService: null,
+	mirrorLodging: null,
+	lodgingSelect: null,
+	loading: false,
+	editMode: false,
+	lodgings: new DataSet([]),
+	periods: new DataSet([]),
+	Places: [],
+	lodgingPersons: [],
+	place: null,
+	rangeDatePayments: {},
+	lodgingsPlace: [],
+	countLogingsPlace: 0,
+	rangeDate: {
+		start: null,
+		end: null,
+	},
+	bottomSheet: false,
+};
+
+export default state;
